Added user_id field to functions, now works correctly. Created Github Actions workflow to build images on push

+++ services/api-gateway/handler/media.go
@@ -9,6 +9,7 @@ import (
 	"path/filepath"
 
 	"github.com/chgrape/storage-app/services/api-gateway/clients"
+	"github.com/chgrape/storage-app/services/api-gateway/middleware"
 	pb "github.com/chgrape/storage-app/shared/gen"
 )
 
@@ -24,8 +25,13 @@ func NewMediaHandler(media *clients.MediaClient) *mediaHandler {
 
 func (h *mediaHandler) Download(w http.ResponseWriter, r *http.Request) {
 	id := r.PathValue("id")
+	userID, ok := r.Context().Value(middleware.UserIDKey).(string)
+	if !ok {
+		http.Error(w, "unauthorized", http.StatusUnauthorized)
+		return
+	}
 
-	stream, err := h.media.Client.Download(r.Context(), &pb.DownloadRequest{Id: id})
+	stream, err := h.media.Client.Download(r.Context(), &pb.DownloadRequest{Id: id, UserId: userID})
 	if err != nil {
 		http.Error(w, "file not found", http.StatusNotFound)
 		return
@@ -48,7 +54,13 @@ func (h *mediaHandler) Download(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *mediaHandler) List(w http.ResponseWriter, r *http.Request) {
-	res, err := h.media.Client.List(r.Context(), &pb.ListRequest{})
+	userID, ok := r.Context().Value(middleware.UserIDKey).(string)
+	if !ok {
+		http.Error(w, "unauthorized", http.StatusUnauthorized)
+		return
+	}
+
+	res, err := h.media.Client.List(r.Context(), &pb.ListRequest{UserId: userID})
 	if err != nil {
 		http.Error(w, fmt.Sprintf("couldn't fetch records: %v", err), http.StatusInternalServerError)
 		return
@@ -59,6 +71,12 @@ func (h *mediaHandler) List(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *mediaHandler) Upload(w http.ResponseWriter, r *http.Request) {
+	userID, ok := r.Context().Value(middleware.UserIDKey).(string)
+	if !ok {
+		http.Error(w, "unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	file, header, err := r.FormFile("file")
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -80,6 +98,7 @@ func (h *mediaHandler) Upload(w http.ResponseWriter, r *http.Request) {
 		Mime:     mime,
 		Size:     header.Size,
 		Data:     data,
+		UserId:   userID,
 	}
 
 	res, err := h.media.Client.Upload(r.Context(), req)
@@ -94,9 +113,15 @@ func (h *mediaHandler) Upload(w http.ResponseWriter, r *http.Request) {
 
 func (h *mediaHandler) Erase(w http.ResponseWriter, r *http.Request) {
 	uuid := r.PathValue("id")
+	userID, ok := r.Context().Value(middleware.UserIDKey).(string)
+	if !ok {
+		http.Error(w, "unauthorized", http.StatusUnauthorized)
+		return
+	}
 
 	res, err := h.media.Client.Delete(r.Context(), &pb.DeleteRequest{
-		Id: uuid,
+		Id:     uuid,
+		UserId: userID,
 	})
 	if err != nil {
 		http.Error(w, fmt.Sprintf("error: error deleting file: %v", err), http.StatusInternalServerError)
